### 2021-01-19 #2971 @whedon
[PRE REVIEW]: DataCleaningTool: An open-source MATLAB app for cooperative data cleaning
**Submitting author:** @devosmitachatterjee2018 (<a href="http://orcid.org/0000-0001-8041-8391">DEVOSMITA CHATTERJEE</a>)
**Repository:** <a href="https://github.com/devosmitachatterjee2018/DataCleaningTool" target ="_blank">https://github.com/devosmitachatterjee2018/DataCleaningTool</a>
**Version:** v1.0
