### 2021-01-19 #2971 @kthyng
Hi @devosmitachatterjee2018 and thanks for your submission. I've flagged this for a scope query to make sure that the software fits JOSS scope guidelines. I'm not super familiar with MATLAB app details so I will be looking for information from my fellow JOSS editors. This process should take 1-2 weeks. Thanks.
