+++ paper/paper.bib
@@ -1,10 +1,16 @@
-@article{Stekhoven:2012,
-  title={MissForest - non-parametric missing value imputation for mixed-type data},
-  author={Daniel J. Stekhoven and Peter B{\"u}hlmann},
-  journal={Bioinformatics},
-  year={2012},
-  volume={28 1},
-  pages={112-8}
+@article{Stekhoven_2011,
+   title={MissForest--non-parametric missing value imputation for mixed-type data},
+   volume={28},
+   ISSN={1460-2059},
+   url={http://dx.doi.org/10.1093/bioinformatics/btr597},
+   DOI={10.1093/bioinformatics/btr597},
+   number={1},
+   journal={Bioinformatics},
+   publisher={Oxford University Press (OUP)},
+   author={Stekhoven, D. J. and Buhlmann, P.},
+   year={2011},
+   month={Oct},
+   pages={112–118}
 }
   
 @software{Olson:2016,
