### 2021-01-19 #2971 @devosmitachatterjee2018
> 10.1093/bioinformatics/btr597

I have corrected the MISSING DOI issue in paper.bib file.
