+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@article{Stekhoven_2011,
+@article{Stekhoven:2011,
    title={MissForest--non-parametric missing value imputation for mixed-type data},
    volume={28},
    ISSN={1460-2059},
