### 2021-01-19 #2971 @devosmitachatterjee2018
I have corrected the MISSING DOI issue in paper.bib file and checked the generated pdf.
