### 2021-01-26 #2971 @danielskatz
Thanks @galessiorob  - you can do the next step yourself in the future, by using either command I use or ```@whedon assign me as editor```
