### 2021-02-02 #2971 @kthyng
@devosmitachatterjee2018 Actually it's preferable to do it the way you did, so thanks! That way we help keep the pings to people who may not end up reviewing to a minimum.

@galessiorob See above, we do try to not ping people in these issues unless necessary since it can add to their email pile quite a bit.
