### 2021-08-30 #2971 @arfon
@devosmitachatterjee2018 – 👋 apologies this submission has taken so long to move forward. I'm going to help our @galessiorob by taking on this submission for her.

Before I look for a second editor I just wanted to check a couple of things:

- @devosmitachatterjee2018 - are you still interested in pursuing this publication in JOSS?
- @hakanfriden - are you still able to review this submission for us? I realize it has been quite some time since you agreed to review for us 😄
