### 2021-08-30 #2971 @devosmitachatterjee2018
@arfon

Thanks for the update.
Yes I am very much interested to publish our software in JOSS.

Best regards
Devosmita
