### 2021-09-01 #2971 @hakanfriden
Dear ???

Maybe I am not qualified for review as I do not have access to all mentioned toolboxes.

Anyway I could run the program with my current licenses.

/Håkan

From: Devosmita Chatterjee ***@***.***
Sent: Tuesday, 31 August 2021 20:24
To: openjournals/joss-reviews
Cc: hakanfriden; Mention
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: DataCleaningTool: An open-source MATLAB app for cooperative data cleaning (#2971)

@arfon <https://github.com/arfon>

Thanks for the update.
I will search for reviewers within my professional network who have access to a MATLAB2018 (or greater) license and the required toolboxes. I will get back to you as soon as possible.

Best regards
Devosmita
