### 2021-09-01 #2971 @devosmitachatterjee2018
@arfon

Today I talked to Shubo Chakrabarti (Science Gateways Strategist at MathWorks) . He informed that any member of any of the Universities offering Campus Wide Licenses worldwide has access to all toolboxes. There are platforms such as wholetale.org that offer free MATLAB access to anyone precisely for cases such as these. Any reviewer can also download free 30 day trials for any toolbox - https://www.mathworks.com/products/sysid.html or https://www.mathworks.com/products/finance.html.

He would appreciate to talk to you. His Github account is @ShuboChakrabarti.

Best regards
Devosmita
