### 2021-09-01 #2971 @ShuboChakrabarti
Hi Arfon,

As Devosmita mentioned,

1. The easiest way for a reviewer to access MATLAB toolboxes would be to
download a free trial valid for 30 days.
2. If the author so wishes, they might want to upload their code onto a
reproducibility platform such as CodeOcean or WholeTale who we partner with
to offer MATLAB access. This means academics can access MATLAB at those
sites and reproduce code written in MATLAB without a license.
3. There are also about 2000 Universities worldwide who have Campus Wide
Licenses which means anyone affiliated to these Universities have access to
all toolboxes. If a reviewer does not know if they have access to such a
Campus Wide License, they can find out here
<https://www.mathworks.com/academia/tah-support-program/eligibility.html>.
Many might not know about this and would have to contact their license
administrators or IT dept at their schools to find out required steps which
may vary from school to school.

If there are further questions on your end or ideas about making the Open
Science experience for MATLAB users better, please do reach out at my work
email ***@***.***

Best,
Shubo

On Wed, Sep 1, 2021 at 3:10 PM Devosmita Chatterjee <
***@***.***> wrote:

> @arfon <https://github.com/arfon>
>
> Today I talked to Shubo Chakrabarti (Science Gateways Strategist at
> MathWorks) . He informed that any member of any of the Universities
> offering Campus Wide Licenses worldwide has access to all toolboxes. There
> are platforms such as wholetale.org that offer free MATLAB access to
> anyone precisely for cases such as these. Any reviewer can also download
> free 30 day trials for any toolbox -
> https://www.mathworks.com/products/sysid.html or
> https://www.mathworks.com/products/finance.html.
>
> He would appreciate to talk to you. His Github account is
> @ShuboChakrabarti <https://github.com/ShuboChakrabarti>.
>
> Best regards
> Devosmita
>
