### 2021-09-01 #3680 @whedon
[REVIEW]: DataCleaningTool: An open-source MATLAB app for cooperative data cleaning
**Submitting author:** @devosmitachatterjee2018 (<a href="http://orcid.org/0000-0001-8041-8391">DEVOSMITA CHATTERJEE</a>)
**Repository:** https://github.com/devosmitachatterjee2018/DataCleaningTool
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
