### 2021-11-03 #3680 @devosmitachatterjee2018
@arfon

I will be grateful if you give me little more time. I am extremely interested to publish our software in JOSS.

1. I already received some comments from the reviewer in a private email.
2. I worked on them and sent the rebuttal to the reviewer.
3. If you wish, I can attach the emails here.
4. I thought that I will update the github repository with the latest version of my software once he give his feedback in the review section.
5. I am trying to contact the reviewer.

Best regards
Devosmita
