### 2021-11-03 #3680 @devosmitachatterjee2018
@arfon

On Sep 1,2020, I sent this email to the reviewer.

Dear Hakan,

I started working on your following feedback:-
1. Remove the 'expected error of imputation for numerical and categorical features' part since it is not the same for all datasets.
2. Set the number of bins for histograms.
3. Add constants to transformations such as log10(C1*X + C2).
4. Reverse the row order with respect to a dataset in the missingness plot
5. Set the number of trees in the missForest algorithm since accuracy increases with increase in number of trees.
6. Improve performance of the code and reduce runtime in my code.
I didn't modify (push and commit) my github repository yet since the JOSS submission was on hold.

We would be honoured to have you as our reviewer. Since you already gave me feedback,  I will try my best to work on them and it will be really nice to continue our collaboration. We value your time.

Best regards
Devosmita
