### 2021-11-03 #3680 @devosmitachatterjee2018
@arfon

On Sep 6,2020, I sent this email to the reviewer.

Dear Håkan,

Hope you are good.

I revised our software as per your following feedback:-

1. Remove the 'expected error of imputation for numerical and categorical features' part in the Imputation widget (since it is not the same for all datasets) - Done.

2. Set the number of bins for histograms in the Numerical Features widget and the Data Transformation widget - Done.

3. Reverse the row order with respect to a dataset in the missingness plot in the Data Properties widget and the Imputation widget - Done.

4. Set the number of trees in the missForest algorithm in the Imputation widget (since accuracy increases with increase in number of trees) - Done.

5. Improve performance of the code and reduce runtime in my code - I tried to reduce the warnings marked in yellow.

I didn't update my user manual though. I thought I will do that at the end after all corrections are made.

I am sharing my code with you in GoogleDrive: https://drive.google.com/drive/folders/1aAWoT33w17XFEOXtAYGx4yQf-ReoqwSS?usp=sharing.

Please let me know if you have any query.

Thank you for your valuable time and guidance.
Best Regards
Devosmita
Ph No: +46 709669519
