### 2021-11-03 #3680 @devosmitachatterjee2018
@arfon

On Sep 8,2020, I received this email from the reviewer.

Hej Devosmita!

I have run your last version.

It is still a bit slow in performance,

I have very hard to accept your imputation algorithm.

Could you describe its procedure in a way that I could grasp?

How on earth could you predict inhabitants or longitude/latitude from these data?

I have 3 days off now, back in business next week, will do my reporting then.

BR
/Håkan
