### 2021-11-03 #3680 @devosmitachatterjee2018
@arfon

On Sep 13,2020, I sent this email to the reviewer.

Dear Håkan,

I hope you have a wonderful weekend.

I have these following things to convey to you.

I.  It is still a bit slow in performance. -
I agree.
1. Matlab app designer has limitations of its own.
2. As suggested, I have tried to solve the warnings (marked in yellow) in the app designer script as much as possible.
3. I will try to break down the app designer script into some functions, write them in separate .m files, and use function calling in the app designer script, and see if the performance increases.

II.  Could you describe its procedure in a way that I could grasp? -
MissForest imputation method is an imputation method based on Random forests.
The procedure by which the Random forests algorithm works.
1. It is an ensemble learning method which means, a method that creates multiple models and then combines them to produce improved results.
2. It comprises a large number of decision trees and makes predictions over categorical or numerical response variables.
3. It outputs the class that is the mode/majority of the predicted classes (for categorical variables) or the mean prediction of the individual trees (for numerical variables).
The below figure (from my Master Thesis Report) shows the working mechanism of the Random forests method.
image.png
The procedure by which the MissForest algorithm works.
1. It is an iterative imputation method that is based on Random forests.
2. It trains Random forests on the observed data, followed by predicting the missing data.
The below figure (from my Master Thesis Report) shows the working mechanism of MissForest imputation method.
pic.png
We use the MissForest imputation method in our data cleaning tool because of the following advantages.
1. The MissForest method allows missing value imputation on both numerical and categorical variables simultaneously.
2. MissForest methods do not require tuning of parameters such as standardization of numerical variables or dummy coding of categorical variables.
3. The MissForest method can be applied to high dimensional datasets.
4. The MissForest method can handle large amounts of missing data.

III. How on earth could you predict inhabitants or longitude/latitude from these data? -
I understand.
1. Our example dataset may have certain limitations to demonstrate the strength of the MissForest imputation method.
2. Example: Prediction of 'Population_Size' (numerical feature) depends on 'Lockdown_Type', and 'Tourism'.
    (i) If 'Population_Size' is large, then 'Lockdown_Type' is "Full" because normally full lockdown is compulsory for a densely populated country.
   (ii) 'Population_Size' should be larger than 'Tourism' (tourism size) because the size of population should normally be larger than the number of tourists.
[A similar example (missing value imputation of 'Age' based on 'Score' using MissForest method) is shown in https://towardsdatascience.com/missforest-the-best-missing-data-imputation-algorithm-4d01182aed3.]
3. Example: Prediction of 'Longitude'/ 'Latitude' (both numerical features) depends on  'Population_Size', 'Tourism', and 'Lockdown_Type'.
   (i) Location data such as 'Longitude'/ 'Latitude' is around the North Pole and the South Pole if 'Population_Size' is small because the regions near the North Pole and the South Pole have comparatively less population.
  (ii) 'Longitude'/ 'Latitude' is not near the North Pole and the South Pole if 'Tourism' is large because the regions near the North Pole and the South Pole normally have less tourists.
 (iii) 'Longitude'/ 'Latitude' is in the middle of Earth if  'Lockdown_Type'  is 'Full' because the regions in the middle of Earth which are densely populated normally should have full lockdown.
[A similar example is shown in the article Random Forest location prediction from social networks during disaster events in https://hal.archives-ouvertes.fr/hal-02306003/document.]

Please let me know your feedback.
Thank you very much for your time and guidance.

Best Regards,
Devosmita
