### 2021-11-03 #3680 @arfon
> I will be grateful if you give me little more time. I am extremely interested to publish our software in JOSS.

OK thanks. I'm happy to continue, especially as you've been receiving feedback from @hakanfriden via email.

@hakanfriden – JOSS is an open peer review journal, as such, we would strongly prefer that you leave this feedback _here on this thread_ or as issues on the software being reviewed (https://github.com/devosmitachatterjee2018/DataCleaningTool). Also, we ask that you please update the checklist as you work through your review.

If you have any questions about the review mechanics (docs [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)), don't hesitate to ask!
