### 2021-11-22 #3680 @devosmitachatterjee2018
@arfon I received this email from the reviewer on Nov 5.

Dear Devosmita!

A quick reflection:

I read the link you gave me: https://towardsdatascience.com/missforest-the-best-missing-data-imputation-algorithm-4d01182aed3

The imputation in the example given assumes Age is correlated to Score. How can they assume this?

In General: How can you assume that other variables in a dataset are correlated to variables with missing values? Is there any check for this? Does the algorithm check for these correlations? (It goes back to the reasoning of long/lat, and what you wrote about countries close to the pole have smaller Population Size. I would say not necessarily, as you would at least have to adjust for country area size)

How do you validate the success/reliability of the imputations?

Do you get any confidence interval for the imputed (predicted) missing values?

BR
/Håkan
