### 2021-11-22 #3680 @devosmitachatterjee2018
@arfon I sent this email to the reviewer on Nov 17.

Dear Håkan,

My answers to your questions are as follows.

1. I read the link you gave me: https://towardsdatascience.com/missforest-the-best-missing-data-imputation-algorithm-4d01182aed3. The imputation in the example given assumes Age is correlated to Score. How can they assume this?

The given link assumes that Age is correlated to Score using logical reasoning which is called data interpretation.

Mathematically, as per literature (https://en.wikipedia.org/wiki/Intelligence_quotient#:~:text=Historically%2C%20IQ%20was%20a%20score,to%20obtain%20the%20IQ%20score.), Intelligence Quotient (IQ) score is highly correlated to the Chronological Age (CA) and the Mental Age (MA) of a person. The formula is given by IQ=MA/CA * 100.

2. In General: How can you assume that other variables in a dataset are correlated to variables with missing values? Is there any check for this?

MissForest algorithm is an imputation method based on random forests which trains a random forest on the observed data/values, followed by predicting the missing data/values using the trained random forest. Please refer to section 2, page 113 of the attached paper.

3. Does the algorithm check for these correlations? (It goes back to the reasoning of long/lat, and what you wrote about countries close to the pole have smaller Population Size. I would say not necessarily, as you would at least have to adjust for country area size).

As per my understanding, correlation and regression are two different terms. Correlation determines the strength between two variables while regression determines how one variable affects the other.

Since a random forest comprises a number of regression tasks, the algorithm focuses on

Regression relationship between the response and the predictor.
Correlation between predictors to determine multicollinearity.
Random Forest uses bootstrap sampling (row sampling) and feature sampling (column sampling). Therefore, Random Forest is not that much affected by multicollinearity/correlations since it picks random sample subsets for different models and thus every model sees a different set of data points.

4. How do you validate the success/reliability of the imputations?

The success/reliability of the imputations can be validated only if we have the true dataset by using the formula

Normalized Root Squared Mean Error for numerical variables.
Percentage of Erroneous categorical entries for categorical variables.
![image](https://user-images.githubusercontent.com/49337890/142881599-c39891b6-e507-42d6-9580-23a54bbc4992.png)

5. Do you get any confidence interval for the imputed (predicted) missing values?

We do not get any confidence interval for the amount of missing values as such. But the performance analysis of the missForest algorithm for different percentages of missing values for continuous data, categorical data and mixed-type data is analysed in the following sections.
Please refer to Fig.1-3, pages 115-116 of the attached [paper](https://academic.oup.com/bioinformatics/article/28/1/112/219101).
Please refer to section 4.1 (Performance Analysis of the MissForest Method), page 45 of the attached Master [Thesis](https://odr.chalmers.se/handle/20.500.12380/302324).

Please let me know if you have any queries.

Thank you very much.
Best regards
Devosmita
