### 2021-11-22 #3680 @devosmitachatterjee2018
@arfon I received this email from the reviewer on Nov 5.

Dear Devosmita!

In regression modelling, X -> Y, there is a procedure called cross-validation.
In essence it implies iteratively holding out parts of a dataset, often 1/7, making a model on the remaining part, using the X of the held-out part to predict its Y.
Since we know the held-out Y, we can evaluate the predictability of these Ys. We get a number, model predictability computed over all iterations, often called Q2.
Q2 is always less than R2 (degree of fit)
In PLS, Q2 is used as a stop criteria to avoid overfitting a model by including to many components (parameters).
You know, with enough number of variables you can always get an R2=1.0 in regression.

We would know that we have an overfitted model if Q2 << R2!
We would know that the model is weak if R2 is low.

What are the goodness values of the imputation procedure you use?

/Håkan
