### 2021-11-22 #3680 @hakanfriden
Hello Devosmita!

I am still very suspicious to the used imputation procedure.
As an experiment I created a uniformly distributed random dataset [50*50], RndMx.xlsx
[ImputationEvaluation.xlsx](https://github.com/openjournals/joss-reviews/files/7584075/ImputationEvaluation.xlsx)
[RndMx.xlsx](https://github.com/openjournals/joss-reviews/files/7584076/RndMx.xlsx)
.
I replaced 10 (20%) of the values in the first variable by NaN, and exported the resulting dataset into an Excel file.
Then I imported the dataset, and imputed the missing values, and saved the imputation errors in the attached file.
There I computed S2 = -0.18, quite bad, and NRMSE = 1.07.
You can check the results in attached ImputationEvaluation.xlsx

I think, as I suspected, the Imputation procedure severely fails on these data, and there is no warning nor message that the imputation results are suspicious. Or, do I miss something?
My subsequent conclusion is that the imputation success depends on the included variables being correlated.
How do you evaluate the validity of the imputations?

Warm regards from Nerja, Spain
/Håkan
