### 2021-11-22 #3680 @arfon
Hi @hakanfriden – thanks for providing your feedback here! When you get a chance, please update your review checklist at the [top of this issue ](https://github.com/openjournals/joss-reviews/issues/3680#issue-985209930) so we can get an idea of roughly how much of your review you have completed thus far.

Many thanks!
