### 2021-11-23 #3680 @hakanfriden
Dear Devosmita

What about the bug I encountered?
Now, the program crashes during the Imputation procedure.
We could look at it on a Teams session if you with.

/Håkan
