### 2021-11-24 #3680 @devosmitachatterjee2018
@hakanfriden

I tried to check the bug (that you encountered) today with the latest version of my software and the Impute button worked for for your dataset 'RndMx.xlsx'.

Before imputation
![image](https://user-images.githubusercontent.com/49337890/143308584-6891a250-5db8-4201-bdd6-5ec635bea5b6.png)

After imputation
![image](https://user-images.githubusercontent.com/49337890/143309533-4488af5f-4a74-4bbf-a5d2-377455b71421.png)

Best regards
Devosmita
