### 2021-11-25 #3680 @devosmitachatterjee2018
> Hello Devosmita!
>
> I am still very suspicious to the used imputation procedure. As an experiment I created a uniformly distributed random dataset [50*50], RndMx.xlsx [ImputationEvaluation.xlsx](https://github.com/openjournals/joss-reviews/files/7584075/ImputationEvaluation.xlsx) [RndMx.xlsx](https://github.com/openjournals/joss-reviews/files/7584076/RndMx.xlsx) . I replaced 10 (20%) of the values in the first variable by NaN, and exported the resulting dataset into an Excel file. Then I imported the dataset, and imputed the missing values, and saved the imputation errors in the attached file. There I computed S2 = -0.18, quite bad, and NRMSE = 1.07. You can check the results in attached ImputationEvaluation.xlsx
>
> I think, as I suspected, the Imputation procedure severely fails on these data, and there is no warning nor message that the imputation results are suspicious. Or, do I miss something? My subsequent conclusion is that the imputation success depends on the included variables being correlated. How do you evaluate the validity of the imputations?
>
> Warm regards from Nerja, Spain /Håkan

@hakanfriden

I checked the above calculations as follows.

1. I deleted the first variable 'ONam' from the matrix [RndMx.xlsx](https://github.com/openjournals/joss-reviews/files/7584076/RndMx.xlsx) for simplicity. Now we have a 50*50 random matrix with 10 missing values.

2. Use the following true matrix and imputed matrix.
[RndMx_Imputed.xlsx](https://github.com/openjournals/joss-reviews/files/7602271/RndMx_Imputed.xlsx)
[RndMx_True.xlsx](https://github.com/openjournals/joss-reviews/files/7602272/RndMx_True.xlsx)

3. Use the following formula for NRSME.
![image](https://user-images.githubusercontent.com/49337890/143425020-e636ec72-8242-4a6e-981d-3fd9945f692b.png)
We see here that the variance is calculated on the whole true matrix (not just the true values).

4. Use the following Matlab code.
>> fullname = 'C:\Users\Acer\Desktop\DataCleaningTool_revised\RndMx_True.xlsx';opts = detectImportOptions(fullname,'NumHeaderLines',0);Tbl_True = readtable(fullname,opts);
>> fullname = 'C:\Users\Acer\Desktop\DataCleaningTool_revised\RndMx_Imputed.xlsx';opts = detectImportOptions(fullname,'NumHeaderLines',0);Tbl_Imputed = readtable(fullname,opts);
>> NRSME = mean(sqrt(mean((Tbl_True{:,:} - Tbl_Imputed{:,:}).^2)./var(Tbl_True{:,:})));

5. I get the following NRSME for the whole dataset.
NRSME = 0.01

Best regards
Devosmita
