### 2021-11-25 #3680 @devosmitachatterjee2018
> Hello again!
>
> I get an error when imputing: Error using DataCleaningTool/ImputeButtonPushed (line 4472) Undefined function 'eq' for input arguments of type 'cell'.
>
> Error using matlab.ui.control.internal.controller.ComponentController/executeUserCallback (line 410) Error while evaluating Button PrivateButtonPushedFcn. The error occurs on line 4472: writetable(Tbl_true,'filesize.csv') In classdef DataCleaningTool
>
> By the way am I using your last version? Where do I find the software version?

@hakanfriden

Please find the latest version of the software (updated code) via the repository link:
[https://github.com/devosmitachatterjee2018/DataCleaningTool/tree/main/Functions](https://github.com/devosmitachatterjee2018/DataCleaningTool/tree/main/Functions)

Please let me know if you have any query.
Best regards
Devosmita
