### 2021-11-28 #3680 @hakanfriden
Hi Devosmita!
I was so puzzled to see you NRSME of 0.1!
Today I read your message more accurately.
The dataset has 50*50=2500 values, I had excluded 10 values in the first column.
When I looked at your formula I see that you made the sum over all 2500 values, so even those that were not NaN-ed, and thus have no error!
For the computation, you should **only sum over the excluded** 10 values, as I showed you in my message., as you can see in the file I recently attached, ImputationEvaluation.xlsx.
Warm regards, Håkan
