### 2021-11-29 #3680 @devosmitachatterjee2018
> Hi Devosmita! I was so puzzled to see you NRSME of 0.1! Today I read your message more accurately. The dataset has 50*50=2500 values, I had excluded 10 values in the first column. When I looked at your formula I see that you made the sum over all 2500 values, so even those that were not NaN-ed, and thus have no error! For the computation, you should **only sum over the excluded** 10 values, as I showed you in my message., as you can see in the file I recently attached, ImputationEvaluation.xlsx. Warm regards, Håkan

@hakanfriden

Thank you for your reply.

I started checking the calculations and soon will upload my answer and the corrections.

Best regards
Devosmita
