### 2021-12-01 #3680 @devosmitachatterjee2018
> Hi Devosmita! I was so puzzled to see you NRSME of 0.1! Today I read your message more accurately. The dataset has 50*50=2500 values, I had excluded 10 values in the first column. When I looked at your formula I see that you made the sum over all 2500 values, so even those that were not NaN-ed, and thus have no error! For the computation, you should **only sum over the excluded** 10 values, as I showed you in my message., as you can see in the file I recently attached, ImputationEvaluation.xlsx. Warm regards, Håkan

@hakanfriden

1. Yes I agree with your point that that the NRMSE should be calculated only over the missing entries of the true matrix and the imputed matrix.
2. MissForest imputation method is an imputation method based on random forests which takes into account the correlation between the explanatory variables. The random data you used is RndMx.xlsx. This data consists of random elements which are uncorrelated. Therefore, we get the NRMSE close to one which means that the network does not perform better than a random prediction.
3. I used the well known Iris dataset (which contains correlated explanatory variables/variables depend on one another) to illustrate the missForest imputation method in the following way.
- Firstly, I use the Matlab code to extract the true matrix (X_True) and the missing matrix (X) in a desired folder.

% Work with the iris data
load fisheriris.mat
A=array2table(meas);
B=cell2table(species);
X_True = horzcat(A,B);% table
X_True.Properties.VariableNames(1:5) = {'Sepal_Length','Sepal_Width','Petal_Length','Petal_Width','Species'};

% Produce the missing data
X1  = ones(size(X_True));
matrix_size = numel(X1);
missingNumber = round(0.1*matrix_size);
X1(randperm(matrix_size, missingNumber))= missing;

array_numeric = varfun(@isnumeric,X_True,'output','uniform');
index_numeric = find(array_numeric==1);
index_text = find(array_numeric==0);

X = X_True;

for i = 1:length(index_numeric)
    indices = find(isnan(X1(:,index_numeric(i)))==1);
    for k = 1:length(indices)
        X{indices(k),index_numeric(i)} = NaN;
    end
end

for i2 = 1:length(index_text)
    indices2 = find(isnan(X1(:,index_text(i2)))==1);
    for k2 = 1:length(indices2)
        X{indices2(k2),index_text(i2)} = {''};
    end
end

writetable(X,'C:\Users\Acer\Desktop\DataCleaningTool_revised\Example\X.csv')
writetable(X_True,'C:\Users\Acer\Desktop\DataCleaningTool_revised\Example\X_True.csv')

- Secondly, I use the DataCleaningTool app to impute the missing values in X and extract the imputed matrix (X_Imputed).

- Then, I use the Matlab code to find Normalised Root Squared Mean Error (NRSME) for continuous/numerical variables and Percentage of Erroneous Categorical entries (PEC) for categorical variables.

% Load data into table
fullname = 'C:\Users\Acer\Desktop\DataCleaningTool_revised\Example\X.csv';opts = detectImportOptions(fullname,'NumHeaderLines',0);Tbl = readtable(fullname,opts);
fullname = 'C:\Users\Acer\Desktop\DataCleaningTool_revised\Example\X_True.csv';opts = detectImportOptions(fullname,'NumHeaderLines',0);Tbl_True = readtable(fullname,opts);
fullname = 'C:\Users\Acer\Desktop\DataCleaningTool_revised\Example\X_Imputed.csv';opts = detectImportOptions(fullname,'NumHeaderLines',0);Tbl_Imputed = readtable(fullname,opts);

% Work with arrays
array_numeric = varfun(@isnumeric,Tbl_True,'output','uniform');
index_numeric = find(array_numeric==1);
index_text = find(array_numeric==0);
X_text = Tbl_True(:,index_text);

X = table2array(Tbl(:,index_numeric));
X_True = table2array(Tbl_True(:,index_numeric));
X_Imputed = table2array(Tbl_Imputed(:,index_numeric));

% Data normalization
X = rescale(X,'InputMin',min(X),'InputMax',max(X));
X_True = rescale(X_True,'InputMin',min(X_True),'InputMax',max(X_True));
X_Imputed = rescale(X_Imputed,'InputMin',min(X_Imputed),'InputMax',max(X_Imputed));

% Calculate NRMSE
indices_nan = find(isnan(X)==1);
NRSME = sqrt(mean((X_True(indices_nan) - X_Imputed(indices_nan)).^2))/std(X_True(indices_nan))

% Calculate PEC
I = 0;
for j = 1:length(index_text)
    j_text = index_text(j);
    for i = 1:size(Tbl,1)
        if isequal(Tbl_Imputed{i,j_text}, Tbl_True{i,j_text}) == 0
            I = I+1;
        end
    end
end
PEC = I/numel(X_text)

- Finally, I got approximately NRMSE ~ 0.12, PEC ~ 0.01. Here we see that the NRMSE is close to 0 since the variables depend on one another.

Please let me know if you have any query.
Best regards
Devosmita
