### 2021-12-02 #3680 @hakanfriden
Hi, just saw your mail.
So we seem to agree that the success of the imputation depends on the correlation between the variable.
I am still very suspicious of the method, as I you give no indication of the reliability of the imputation result, or do you?
If there are missing values in a variable that has little correlation with the rest, the results are very dubious!

My preferred way of handling NaNs is to use methods that can handle NaNs, such as PCA and PLS.

I will run your new program again as I get the time for it.

/Håkan
