### 2021-12-02 #3680 @devosmitachatterjee2018
> Hi, just saw your mail. So we seem to agree that the success of the imputation depends on the correlation between the variable. I am still very suspicious of the method, as I you give no indication of the reliability of the imputation result, or do you? If there are missing values in a variable that has little correlation with the rest, the results are very dubious!
>
> My preferred way of handling NaNs is to use methods that can handle NaNs, such as PCA and PLS.
>
> I will run your new program again as I get the time for it.
>
> /Håkan

@hakanfriden

Thank you for your reply.

1. So we seem to agree that the success of the imputation depends on the correlation between the variable.

- Yes I agree that the missForest imputation method depends on the correlation between the variables. The reason is that the imputation method is based on Random Forests that uses observed values in each variable as the dependent variable and all the other variables in the data as the independent variables to impute missing values for each variable, until convergence.

2. I am still very suspicious of the method, as you give no indication of the reliability of the imputation result, or do you? If there are missing values in a variable that has little correlation with the rest, the results are very dubious!

- Given that the variables depend on one another/are correlated, the reliability of the imputation result depends on the Normalised Root Squared Mean Error (NRSME) for continuous/numerical variables and the Percentage of Erroneous Categorical entries (PEC) for categorical variables. Therefore, the NRMSE and the PEC values should be close to zero. [In the previous message, I used the Iris dataset (correlation between the variables) to illustrate the missForest imputation method. For 10% missing data, I got approximately the NRMSE value ~ 0.12 and the PEC value ~ 0.01 .]

Otherwise, if the the variables are uncorrelated/less correlated, then the imputation method gives just random prediction which is inaccurate.

3. My preferred way of handling NaNs is to use methods that can handle NaNs, such as PCA and PLS.

- The missForest imputation method is implemented in our tool for imputing continuous as well as categorical variables simultaneously without data transformation.

Please let me know if you have any query.
Best regards
Devosmita
