### 2021-12-13 #3680 @hakanfriden
Hi Devosmita,

Might you be available this week for a discussion?
I downloaded your latest application, and installed it.
I have a dataset of around 60000 obs by some 50 variables.
Could we have a discussion over Teams, and maybe look at some other datasets.
Maybe you have a suitable dataset as well?

I have a time slot tomorrow 11-12, and any time after 19
Wednesday from 13 and on.
Thursday 10-13, any time after 14:30
Friday 10-13, 14-17.

/Håkan
