### 2021-12-14 #3680 @devosmitachatterjee2018
@hakanfriden

Thank you for your response.

Could we have the meeting Wednesday (15.12.2021) at 20:00? In that case, I will send you a zoom link before the meeting.

Best regards
Devosmita
