### 2022-02-18 #3680 @arfon
@devosmitachatterjee2018 – it is over 1 year since this submission was first made to JOSS. Given the incredibly slow progress here, and severe challenges finding multiple reviewers, I don't believe this submission is likely to ever actually make it through review at JOSS. Do you disagree?
