### 2022-07-12 #3680 @arfon
@devosmitachatterjee2018 – revisiting this thread from February and my earlier comment: Given the incredibly slow progress here, and severe challenges finding multiple reviewers, I don't believe this submission is likely to ever actually make it through review at JOSS. Do you disagree?
